Made Unicode strings the default and moved PregTool to separate file
