+++ l2tpreg/definitions.py
@@ -1,13 +1,16 @@
 # -*- coding: utf-8 -*-
 """The definitions."""
 
-REGISTRY_FILE_TYPE_NTUSER = u'NTUSER'
-REGISTRY_FILE_TYPE_SAM = u'SAM'
-REGISTRY_FILE_TYPE_SECURITY = u'SECURITY'
-REGISTRY_FILE_TYPE_SOFTWARE = u'SOFTWARE'
-REGISTRY_FILE_TYPE_SYSTEM = u'SYSTEM'
-REGISTRY_FILE_TYPE_UNKNOWN = u'UNKNOWN'
-REGISTRY_FILE_TYPE_USRCLASS = u'USRCLASS'
+from __future__ import unicode_literals
+
+
+REGISTRY_FILE_TYPE_NTUSER = 'NTUSER'
+REGISTRY_FILE_TYPE_SAM = 'SAM'
+REGISTRY_FILE_TYPE_SECURITY = 'SECURITY'
+REGISTRY_FILE_TYPE_SOFTWARE = 'SOFTWARE'
+REGISTRY_FILE_TYPE_SYSTEM = 'SYSTEM'
+REGISTRY_FILE_TYPE_UNKNOWN = 'UNKNOWN'
+REGISTRY_FILE_TYPE_USRCLASS = 'USRCLASS'
 
 REGISTRY_FILE_TYPES = frozenset([
     REGISTRY_FILE_TYPE_NTUSER,
